kaut kas izlabots

+++ csharp-basics/exercises/Polymorphism/AdApp/Hoarding.cs
@@ -14,9 +14,9 @@ namespace AdApp
         {
         }
 
-        public new int Cost() 
+        public new double Cost() 
         {
-            return (int)base.Cost();
+            return base.Cost();
         }
 
         public override string ToString()
